pyproject.toml
[project]
name = "fastapi-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "mysql-connector-python>=9.4.0",
    "nanoid>=2.0.0",
    "ruff>=0.13.1",
]
